Updates dependencies and edition to 2024
Updates the `Cargo.lock` file to reflect updated dependencies and switches the edition to 2024.
This ensures compatibility with the latest features and improvements in the Rust ecosystem.
The `fluxion_macro` crate is now specified using a relative path instead of a registry version.
Additionally updates the slacktor dependency to point to a git repository.